intro work

+++ frontend/build.sh
@@ -1,21 +1,23 @@
-#!/usr/bin/env sh
+#!/usr/bin/env bash
 set -e
 
-pnpm svg
 pnpm build:tsc
 pnpm build:vite
 
-mv dist/index.html dist/original.html
+if [ "$1" == "--minify" ]; then
+    mv dist/index.html dist/original.html
 
-pnpm exec minify-html \
-  --minify-css \
-  --minify-js \
-  --do-not-minify-doctype \
-  --ensure-spec-compliant-unquoted-attribute-values \
-  --keep-spaces-between-attributes \
-  --output dist/index.html \
-  dist/original.html
+    pnpm exec minify-html \
+      --minify-css \
+      --minify-js \
+      --do-not-minify-doctype \
+      --ensure-spec-compliant-unquoted-attribute-values \
+      --keep-spaces-between-attributes \
+      --output dist/index.html \
+      dist/original.html
 
-rm dist/original.html
+    rm dist/original.html
+fi
+
 
 (cd dist && zip -qr ../dist.zip .)
